move cd rendering into component, filter 00 cds
